Implémentation du décompte des symptômes et modification de la signature de la méthode dans la classe Launcher pour passer d'un objet Integer à un objet Long dans la Map retournée

+++ Project02Eclipse/src/com/hemebiotech/analytics/LaunchProgram.java
@@ -12,7 +12,7 @@ public class LaunchProgram {
 
             List<String> symptomsList = symptomReader.GetSymptoms();
 
-            Map<String, Integer> symptomsMap = AnalyticsCounter.CountStringOccurences(symptomsList);
+            Map<String, Long> symptomsMap = AnalyticsCounter.CountStringOccurences(symptomsList);
 
             OutputFileWriter.WriteInFile(symptomsMap);
         }
